Add files via upload
Here is the code ive been working on, the speed calculation part is not in there yet. So, when this code is ran the video will play, when the video is playing press s to  pause the video, and you can use the mouse to click on the video to mark the corners of the box.